Introduce StableStorage trait
* Changes:

- Introduce StableStorage trait.
- Make StableStorage a pre-requisite of PinningStorage.
- Move memory stability guarantees from MultipleStorage to
  StableStorage.
- Adjust AllocatorStorage and SkipList accordingly.

* Motivation:

A conceivably odd form of Storage would be caching storage, fetching the
stored memory from compressed form, or from disk or network, and only
ever keeping a few records in memory at a time.

Such a form of Storage would be a MultipleStorage, but would not
otherwise provide any stability of memory addresses.

While seemingly exotic, it is a good reminder that bundling together
seemingly unrelated guarantees may not be very forward looking. And
since it is easier to add guarantees than to remove them, it seems
better to start with more traits and more fine-grained guarantees.