Introduce SharingStore and LinkedList.
* Changes:

- Introduce SharingStore, implemented for Global.
- Introduce LinkedList.
- Introduce SizedHandle...

* Motivation:

LinkedList::append and LinkedList::split_off can only be implemented if
two different instances of a store can use one-another handle. This
requires a very specific property, which SharingStore makes explicit.

* Limitation:

A weird edge case in the compiler prevented reusing TypedHandle in the
implementation; as a work-around SizedHandle was introduced for
expendiency's sake. It is intended to remain a work-around.