Introduce InlineBumpStore
* Changes:

- Introduce InlineBumpStore.
- Duplicate LinkedList test suite for it -- minus SharingStore tests.

* Motivation:

InlineBumpStore is an example of MultipleStore + StableStore that is not
an allocator.

InlineBumpStore is also a fairly generic store, usable with most
collection types, and quite likely the store to use to create a const or
static collection at compile-time.

InlineBumpStore also allows demonstrating handle compression. In the
LinkedList tests, it uses a single-byte handle, instead of a 4-bytes or
8-bytes pointer. When LinkedList is used with an f32, this means that a
Node takes 8 bytes instead of 24 bytes, a fairly drastic compression.