Introduce AllocatorHandle
* Changes:

- Introduce AllocatorHandle, a Send+Sync wrapper around NonNull<u8>.
- Use AllocatorHandle for Allocators.

* Motivation:

It is possible that handles may not be Send or Sync, in which case a
collection containing such a handle should NOT override this property,
else it be unsound.

In turn, this thus means that for Vec<T, A> to be Send+Sync, as it is
today, then the handle obtained from A must be Send+Sync... and thus a
wrapper around NonNull<u8> is required.