Switch from Store and StoreMultiple to Store and StoreSingle
* Changes:

- "Merge" StoreMultiple into Store.
- Introduce parallel StoreSingle.
- Rework extensions and collections accordingly.

* Motivation:

A store handing out multiple concurrent allocations finds itself facing
an aliasing issue: no reference to a block of memory carved up for
allocations can be formed while any pointer to one of its allocation
exist.

A partial solution is to wrap the block of memory into `UnsafeCell` and
only ever reference it via a shared reference, which in turns lead to
all methods of `Store` taking a `&self`.

One special case arises, though. A `Store` that is not a `StoreMultiple`
and only ever handles a single allocation should have no such worries.
And due to how prevalent `Box`, `Vec`, `VecDeque`, or `HashXxx` are, all
only requiring a single allocation, it makes sense to provide an
alternative API specifically tailored for this corner case.

This API is `StoreSingle`.

* Limitation:

For some reasons, attempting to provide a default implementation of
`StoreSingle` for any type implementing `Store` runs afoul of coherence,
so for now a simple forwarding implementation has been copy-pasted...